Allow merged codebins as option.
This is utterly stupid and a lossy removal of information. Unfortunately, it's what you call a "scene standard" which is to say an idiotic decision that somehow stuck.

Anyways, add a toggle for usage of merged codebins. These should go one level above where text/ro/data are and are named by titleid sans extension.