Add a early-ish prototype ips2pco tool.
This tool takes an IPS or IPS32 patch and creates a .pco (bytecode source) skeleton. This is based on a previously written tool in my libbinmod repository for actually creating/applying IPS patches. The core changes to blib.c/blib.h will make it there soon-ish.

This can be used to automate conversion of IPS patches to a format compatible with corbenik's patcher system. Do note the output files are not ready for use as-is -- you will need to manually edit the name, description, and titleID as needed, as well as actually run it through bytecode_asm.py.