Screeninit is broken
I'm still unsure as to *why*, but I assume something I'm doing is probably causing an ARM11 exception.

To be honest, part of the problem here is that I never should have gone with the whole mixing arm11 and arm9 code thing.

For now, I've temporarily made the drawing code work in 24-bit mode and 32-bit mode via #ifdefs and nerfed screeninit from the code, instead adding the '-i' flag to firmtool.